move to adapters
```release-note
NONE
```
